frontend/requirements.txt
streamlit==1.39.0
httpx==0.27.2
pandas==2.2.3
python-dotenv==1.2.1


